Create FX Functions (#10)
Co-Authored-By: irinazoccolini <71565864+irinazoccolini@users.noreply.github.com>, Watermelanie <46464201+Watermelanie@users.noreply.github.com>, and Ashar Farooq <ashar744@gmail.com>

This PR creates filter functions to modify input sound waves for greater flexibility in the audio that is produced from user input. While the filters do not work perfectly, here are the ideas behind each filter:

- `apply_filter`: This is the base function where we call other filter functions for each note.
- `get_notes`: This separates the audio into note slices and identifies the root note being played.
- `change_volume`: This changes the amplitude of audio samples.
- `stretch_audio`: This changes the duration of the audio.
- `change_pitch`: This changes the pitch of the notes in the audio.
- `add_chords`: This builds major or minor triads upon each note in the audio.

There are some helper functions for some of these filters as well as a bare `overlap_notes` filter. The idea was to take an audio signal and overlap notes in the audio by a certain factor, but this still needs to be implemented. See Issue #9 for more details on improvements.